requirements.txt
attrs==21.4.0
aws-cdk-lib==2.20.0
black==22.3.0
boto3==1.21.42
botocore==1.24.42
cattrs==22.1.0
click==8.1.2
constructs==10.0.119
exceptiongroup==1.0.0rc3
jmespath==1.0.0
jsii==1.56.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
